pyproject.toml
[project]
name = "citing-authors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crossrefapi>=1.6.0",
    "more-itertools>=10.6.0",
    "pandas[excel]>=2.2.3",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
